refactor(headless): split RSI replay guard tests
